add test
and restructure tests in multiple files
