+++ test/unit/tool_util/test_tool_linters_input.py
@@ -1,19 +1,8 @@
 import pytest
 
 from galaxy.tool_util.lint import LintContext
-from galaxy.tool_util.linters import general, inputs, outputs
-from galaxy.tool_util.parser.xml import XmlToolSource
+from galaxy.tool_util.linters import inputs
 from galaxy.util import etree
-
-
-def lint_general(xml_tree, lint_ctx):
-    """Wrap calling of lint_general to provide XmlToolSource argument.
-
-    This allows general.lint_general to be called with the other linters which
-    take an XmlTree as an argument.
-    """
-    tool_source = XmlToolSource(xml_tree)
-    return general.lint_general(tool_source, lint_ctx)
 
 
 NO_SECTIONS_XML = """
@@ -104,23 +93,6 @@ SELECT_OPTION_DEFINITIONS = """
 </tool>
 """
 
-WHITESPACE_IN_VERSIONS_AND_NAMES = """
-<tool name=" BWA Mapper " id="bwa tool" version=" 1.0.1 " is_multi_byte="true" display_interface="true" require_login="true" hidden="true">
-    <description>The BWA Mapper</description>
-    <requirements>
-        <requirement type="package" version=" 1.2.5 "> bwa </requirement>
-    </requirements>
-    <version_command interpreter="python">bwa.py --version</version_command>
-    <inputs>
-        <param name="select_fd_op" type="select">
-            <options from_dataset="xyz"/>
-            <options from_data_table="xyz"/>
-            <option value="x">x</option>
-        </param>
-    </inputs>
-</tool>
-"""
-
 VALIDATOR_INCOMPATIBILITIES = """
 <tool name="BWA Mapper" id="bwa" version="1.0.1" is_multi_byte="true" display_interface="true" require_login="true" hidden="true">
     <description>The BWA Mapper</description>
@@ -134,33 +106,6 @@ VALIDATOR_INCOMPATIBILITIES = """
 </tool>
 """
 
-# check that linter accepts format source for collection elements as means to specify format
-# and that the linter warns if format and format_source are used
-OUTPUTS_COLLECTION_FORMAT_SOURCE = """
-<tool name="BWA Mapper" id="bwa" version="1.0.1" is_multi_byte="true" display_interface="true" require_login="true" hidden="true">
-    <description>The BWA Mapper</description>
-    <version_command interpreter="python">bwa.py --version</version_command>
-    <outputs>
-        <collection name="output_collection" type="paired">
-            <data name="forward" format_source="input_readpair" />
-            <data name="reverse" format_source="input_readpair" format="fastq"/>
-        </collection>
-    </outputs>
-</tool>
-"""
-
-# check that linter does not complain about missing format if from_tool_provided_metadata is used
-OUTPUTS_DISCOVER_TOOL_PROVIDED_METADATA = """
-<tool name="BWA Mapper" id="bwa" version="1.0.1" is_multi_byte="true" display_interface="true" require_login="true" hidden="true">
-    <description>The BWA Mapper</description>
-    <version_command interpreter="python">bwa.py --version</version_command>
-    <outputs>
-        <data name="output">
-            <discover_datasets from_tool_provided_metadata="true"/>
-        </data>
-    </outputs>
-</tool>
-"""
 TESTS = [
     (
         NO_SECTIONS_XML, inputs.lint_inputs,
@@ -211,15 +156,6 @@ TESTS = [
             and len(x.warn_messages) == 0 and len(x.error_messages) == 5
     ),
     (
-        WHITESPACE_IN_VERSIONS_AND_NAMES, lint_general,
-        lambda x:
-            "Tool version contains whitespace, this may cause errors: [ 1.0.1 ]." in x.warn_messages
-            and "Tool name contains whitespace, this may cause errors: [ BWA Mapper ]." in x.warn_messages
-            and "Requirement version contains whitespace, this may cause errors: [ 1.2.5 ]." in x.warn_messages
-            and "Tool ID contains whitespace - this is discouraged: [bwa tool]."
-            and len(x.warn_messages) == 4 and len(x.error_messages) == 0
-    ),
-    (
         VALIDATOR_INCOMPATIBILITIES, inputs.lint_inputs,
         lambda x:
             "Parameter [param_name]: 'in_range' validators are not expected to contain text (found 'TEXT')" in x.warn_messages
@@ -228,17 +164,6 @@ TESTS = [
             and "Parameter [param_name]: attribute 'filename' is incompatible with validator of type 'regex'" in x.error_messages
             and "Parameter [param_name]: 'regex' validators need to define an 'expression' attribute" in x.error_messages
             and len(x.warn_messages) == 1 and len(x.error_messages) == 4
-    ),
-    (
-        OUTPUTS_COLLECTION_FORMAT_SOURCE, outputs.lint_output,
-        lambda x:
-            "Tool data output reverse should use either format_source or format/ext" in x.warn_messages
-            and len(x.warn_messages) == 1 and len(x.error_messages) == 0
-    ),
-    (
-        OUTPUTS_DISCOVER_TOOL_PROVIDED_METADATA, outputs.lint_output,
-        lambda x:
-            len(x.warn_messages) == 0 and len(x.error_messages) == 0
     ),
 ]
 
@@ -249,10 +174,7 @@ TEST_IDS = [
     'select duplicated options',
     'select deprecations',
     'select option definitions',
-    'hazardous whitespace',
     'validator imcompatibilities',
-    'outputs collection static elements with format_source',
-    'outputs discover datatsets with tool provided metadata'
 ]
